requirements.txt
langchain>=0.1.0
pydantic>=2.0.0
PyYAML>=6.0
prometheus-client>=0.16.0
fastapi>=0.95.0
uvicorn>=0.22.0
pylint>=2.17.0
black>=23.3.0
bandit>=1.7.5
jinja2>=3.1.2
rich>=13.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
mypy>=1.3.0
cryptography>=40.0.0
